move functions into helper file
